+++ lib/ui/events2/Event2SetupSurveyPanel.dart
@@ -17,20 +17,25 @@
 import 'package:flutter/material.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/ui/events2/Event2CreatePanel.dart';
+import 'package:illinois/ui/events2/Event2Widgets.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/ui/widgets/RibbonButton.dart';
 import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/model/event2.dart';
 import 'package:rokwire_plugin/model/survey.dart';
+import 'package:rokwire_plugin/service/events2.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/service/surveys.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
 class Event2SetupSurveyPanel extends StatefulWidget {
-  final Event2SurveyDetails? details;
+  final Event2? event;
+  final Event2SurveyDetails? surveyDetails;
 
-  Event2SetupSurveyPanel({Key? key, this.details}) : super(key: key);
+  Event2SetupSurveyPanel({Key? key, this.event, this.surveyDetails}) : super(key: key);
+
+  Event2SurveyDetails? get details => (event?.id != null) ? event?.surveyDetails : surveyDetails;
 
   @override
   State<StatefulWidget> createState() => _Event2SetupSurveyPanelState();
@@ -41,14 +46,23 @@ class _Event2SetupSurveyPanelState extends State<Event2SetupSurveyPanel>  {
   late bool _hasSurvey;
   final TextEditingController _hoursController = TextEditingController();
 
+  late bool _initialHasSurvey;
+  late String _initialHours;
+
   List<Survey>? _surveys;
   Survey? _selectedSurvey;
+  
+  bool _modified = false;
   bool _loadingSurveys = false;
+  bool _updatingSurvey = false;
 
   @override
   void initState() {
-    _hasSurvey = widget.details?.hasSurvey ?? false;
-    _hoursController.text = widget.details?.hoursAfterEvent?.toString() ?? '';
+    _hasSurvey = _initialHasSurvey = widget.details?.hasSurvey ?? false;
+    _hoursController.text = _initialHours = widget.details?.hoursAfterEvent?.toString() ?? '';
+    if (_isEditing) {
+      _hoursController.addListener(_checkModified);
+    }
     _loadSurveys();
     super.initState();
   }
@@ -62,7 +76,7 @@ class _Event2SetupSurveyPanelState extends State<Event2SetupSurveyPanel>  {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: HeaderBar(title: Localization().getStringEx('panel.event2.setup.survey.header.title', 'Event Follow-Up Survey'), onLeading: _onHeaderBack),
+        appBar: _headerBar,
         body: _buildPanelContent(),
         backgroundColor: Styles().colors!.white);
   }
@@ -95,8 +109,9 @@ class _Event2SetupSurveyPanelState extends State<Event2SetupSurveyPanel>  {
       _hasSurvey = !_hasSurvey;
       if (!_hasSurvey) {
         _hoursController.text = '';
-      }
+      }      
     });
+    _checkModified();
   }
 
   // Hours
@@ -174,22 +189,99 @@ class _Event2SetupSurveyPanelState extends State<Event2SetupSurveyPanel>  {
     });
   }
 
-  // Submit
+  // HeaderBar
 
-  void _onHeaderBack() {
-    if (_hasSurvey) {
+  bool get _isEditing => StringUtils.isNotEmpty(widget.event?.id);
 
-      int? hoursNumber = int.tryParse(_hoursController.text);
-      if ((hoursNumber == null) || (hoursNumber < 0)) {
-        AppAlert.showDialogResult(context, Localization().getStringEx('panel.event2.setup.survey.hours.invalid.msg', 'Please, fill valid non-negative number for hours.'));
-        return;
-      }
-      //TBD: Ackolwedge _selectedSurvey
+  PreferredSizeWidget get _headerBar => HeaderBar(
+    title: Localization().getStringEx('panel.event2.setup.survey.header.title', 'Event Follow-Up Survey'),
+    onLeading: _onHeaderBarBack,
+    actions: _headerBarActions,
+  );
 
-      Navigator.of(context).pop(Event2SurveyDetails(hasSurvey: true, hoursAfterEvent: hoursNumber));
+  List<Widget>? get _headerBarActions {
+    if (_updatingSurvey) {
+      return [Event2CreatePanel.buildHeaderBarActionProgress()];
+    }
+    else if (_isEditing && _modified) {
+      return [Event2CreatePanel.buildHeaderBarActionButton(
+        title: Localization().getStringEx('dialog.apply.title', 'Apply'),
+        onTap: _onHeaderBarApply,
+      )];
     }
     else {
+      return null;
+    }
+  }
+
+  void _checkModified() {
+    if (_isEditing && mounted) {
+      
+      bool modified = (_hasSurvey != _initialHasSurvey) ||
+        (_hoursController.text != _initialHours);
+
+      if (_modified != modified) {
+        setState(() {
+          _modified = modified;
+        });
+      }
+    }
+  }
+
+  Event2SurveyDetails _buildSurveyDetails() => Event2SurveyDetails(
+    hasSurvey: _hasSurvey,
+    hoursAfterEvent: Event2CreatePanel.textFieldIntValue(_hoursController),
+  );
+
+  bool _checkSurveyDetails(Event2SurveyDetails surveyDetails) {
+    if ((surveyDetails.hasSurvey ?? false) && ((surveyDetails.hoursAfterEvent == null) || ((surveyDetails.hoursAfterEvent ?? 0) < 0))) {
+      AppAlert.showDialogResult(context, Localization().getStringEx('panel.event2.setup.survey.hours.invalid.msg', 'Please, fill valid non-negative number for hours.'));
+      return false;
+    }
+    return true;
+  }
+
+  void _updateEventSurveyDetails(Event2SurveyDetails? surveyDetails) {
+    if (_isEditing && (_updatingSurvey != true)) {
+      setState(() {
+        _updatingSurvey = true;
+      });
+      Events2().updateEventSurveyDetails(widget.event?.id ?? '', surveyDetails).then((result) {
+        if (mounted) {
+          setState(() {
+            _updatingSurvey = false;
+          });
+        }
+
+        if (result is Event2) {
+          Navigator.of(context).pop(result);
+        }
+        else {
+          Event2Popup.showErrorResult(context, result);
+        }
+
+      });
+    }
+  }
+
+  void _onHeaderBarApply() {
+    Analytics().logSelect(target: 'HeaderBar: Apply');
+    Event2SurveyDetails surveyDetails = _buildSurveyDetails();
+    if (_checkSurveyDetails(surveyDetails)) {
+      _updateEventSurveyDetails(surveyDetails);
+    }
+  }
+
+  void _onHeaderBarBack() {
+    Analytics().logSelect(target: 'HeaderBar: Back');
+    if (_isEditing) {
       Navigator.of(context).pop(null);
+    }
+    else {
+      Event2SurveyDetails surveyDetails = _buildSurveyDetails();
+      if (_checkSurveyDetails(surveyDetails)) {
+        Navigator.of(context).pop(surveyDetails);
+      }
     }
   }
 }